+++ jest/setupExtensions.js
@@ -1,1 +1,1 @@
-import "@date-time/DateRange"
+import "@date-time/DateRangeFormatting"
