+++ lib/utils/CallbackCollection.ts
@@ -1,4 +1,4 @@
-import { uuidString } from "./utils/UUID"
+import { uuidString } from "./UUID"
 
 /**
  * A simple class for doing multi-subscriber pub-sub.
@@ -7,11 +7,11 @@ export class CallbackCollection<Arg> {
   // eslint-disable-next-line func-call-spacing
   private callbacks = new Map<string, (arg: Arg) => void>()
 
-  get count () {
+  get count() {
     return this.callbacks.size
   }
 
-  add (callback: (arg: Arg) => void) {
+  add(callback: (arg: Arg) => void) {
     const id = uuidString()
     this.callbacks.set(id, callback)
     return () => {
@@ -19,7 +19,7 @@ export class CallbackCollection<Arg> {
     }
   }
 
-  send (arg: Arg) {
+  send(arg: Arg) {
     // eslint-disable-next-line n/no-callback-literal
     this.callbacks.forEach((callback) => callback(arg))
   }
